fix: how slash commands, modals and message components are registered and received

+++ internal/database/models.go
@@ -2,8 +2,8 @@ package database
 
 type Guild struct {
 	ID                                                  string `gorm:"primaryKey"`
-	MaxWarningsPerUser                                  int8   `gorm:"default:3"`
-	OnReachMaxWarningsPerUser                           int8   `gorm:"default:1"`
+	MaxWarningsPerUser                                  int    `gorm:"default:3"`
+	OnReachMaxWarningsPerUser                           int    `gorm:"default:1"`
 	SecondsToDeleteMessagesForOnReachMaxWarningsPerUser int    `gorm:"default:0"`
 	RoleToGiveOnReachMaxWarningsPerUser                 string
 	SecondsPunishedUserShouldKeepRoleFor                int `gorm:"default:0"`
@@ -13,5 +13,5 @@ type Guild struct {
 type GuildMember struct {
 	UserId   string `gorm:"primaryKey"`
 	GuildId  string
-	Warnings int8 `gorm:"default:0"`
+	Warnings int `gorm:"default:0"`
 }
